added sorting for topic instances in the REST interface
It is now possible to sort the list of topic instances by the columns
that were represented by attribute types of the topic instances. Sorting
on geometry data is not possible because we don't see any benefit for
this. To achieve a generic read method with sorting capabilities we
created a OpenInfraOrderBy object. This object holds either a
OpenInfraOrderByEnum or a UUID. This led to an update of many resources
that used the old enumeration.