added entity synchronization for new created objects
Adding the system column xmin to our POJOs created a problem when
creating a new object. If we create a new object the entity in the JPA
cache will not include the xmin value because this will be created by
PostgreSQL on the database. So if we created a new object and tried to
read this newly created object we received a NullPointerException while
reading xmin. This problem was solved by adding a em.refresh() on the
currently created object. Additionally the cascade type of several model
classes was updated to do a cascading refresh.